// src/onboarding/domain/business-vertical/enums/vertical.enum.ts
export enum VerticalId {
	BELLEZA_CUIDADO_PERSONAL = "belleza_cuidado_personal",
	FITNESS_BIENESTAR = "fitness_bienestar",
	SALUD_MEDICINA = "salud_medicina",
	COMIDA_BEBIDA = "comida_bebida",
	SERVICIOS_PROFESIONALES = "servicios_profesionales",
	RETAIL_ECOMMERCE = "retail_ecommerce",
	SERVICIOS_HOGAR = "servicios_hogar",
	EDUCACION_FORMACION = "educacion_formacion",
	TURISMO_ENTRETENIMIENTO = "turismo_entretenimiento",
	INMOBILIARIA = "inmobiliaria",
	EVENTOS = "eventos",
	OTRO = "otro",
}